fix:biuldエラーの解消
