+++ next.config.ts
@@ -3,7 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   /* config options here */
   images: {
-    domains: ['thumbnail.image.rakuten.co.jp'], // 楽天のホスト名を追加
+    domains: ["thumbnail.image.rakuten.co.jp"], // 楽天のホスト名を追加
   },
 };
