+++ src/components/top/NewItems/NewItems.test.tsx
@@ -1,18 +1,27 @@
-import { render, screen } from "@testing-library/react";
-import NewItems from "./NewItems";
+import { getNewItems } from "./NewItems";
 
-
-jest.mock("../Item/Item", () =>()=> <div>Mocked Item</div>);
-describe("NewItemsコンポーネントのテスト", () => {
+describe("新着アイテムの取得関数のテスト", () => {
   const mockItemData = [
     { itemName: "Item 1", itemCode: "001", imageUrl: "/", itemPrice: 1000 },
     { itemName: "Item 2", itemCode: "002", imageUrl: "/", itemPrice: 2000 },
   ];
-
-  test("正常にUIが表示されることを確認", () => {
-    render(<NewItems newItems={mockItemData} />);
-    const titleText = screen.getByText("新着アイテム");
-
-    expect(titleText).toBeInTheDocument();
+  test("正常に商品データを返すとことを確認", async () => {
+    global.fetch = jest.fn().mockReturnValueOnce({
+      ok: true,
+      status: 200,
+      json: () => ({ items: mockItemData }),
+    });
+    const itemData = await getNewItems();
+    // console.log(itemData);
+    expect(itemData).toBe(mockItemData);
+  });
+  test("エラー時にnullを返すことを確認", async () => {
+    global.fetch = jest.fn().mockReturnValueOnce({
+      ok: false,
+      status: 500,
+    });
+    const itemData = await getNewItems();
+    // console.log(itemData);
+    expect(itemData).toBeNull();
   });
 });
