+++ src/lib/axios/axios.ts
@@ -8,27 +8,15 @@ export const handleAxiosError = (error: unknown) => {
     if (error.response) {
       console.error("サーバーエラー:", error.response.status);
       console.error("エラーメッセージ:", error.response.data);
-      return NextResponse.json(
-        { message: error.response.data },
-        { status: error.response.status }
-      );
+      return NextResponse.json({ message: error.response.data }, { status: error.response.status });
     } else if (error.request) {
       console.error("リクエストエラー:", error.request);
-      return NextResponse.json(
-        { message: "サーバーが応答しませんでした。" },
-        { status: 504 }
-      );
+      return NextResponse.json({ message: "サーバーが応答しませんでした。" }, { status: 504 });
     } else {
       console.error("リクエスト設定エラー:", error.message);
-      return NextResponse.json(
-        { message: "無効なリクエスト設定。" },
-        { status: 400 }
-      );
+      return NextResponse.json({ message: "無効なリクエスト設定。" }, { status: 400 });
     }
   }
   console.error("予期しないエラー:", error);
-  return NextResponse.json(
-    { message: "サーバー内部でエラーが発生しました。" },
-    { status: 500 }
-  );
+  return NextResponse.json({ message: "サーバー内部でエラーが発生しました。" }, { status: 500 });
 };
